backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
httpx==0.27.2
pymupdf==1.24.9
python-docx==1.1.2
python-multipart==0.0.9
pydantic==2.9.2
